# elearning/e-smart-learn/app/docs/swagger.yaml
definitions:
  apperror.AppError:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      type:
        $ref: '#/definitions/apperror.ErrorType'
    type: object
  apperror.ErrorType:
    enum:
    - not_found
    - bad_request
    - internal_server
    - unauthorized
    - forbidden
    - duplicate_entry
    - validation_error
    type: string
    x-enum-varnames:
    - NotFound
    - BadRequest
    - InternalServer
    - Unauthorized
    - Forbidden
    - DuplicateEntry
    - ValidationError
...
  dto.ApiResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/apperror.AppError'
        type: array
      metadata: {}
      path:
        type: string
      process_id:
        type: string
      request: {}
      status:
        $ref: '#/definitions/dto.ResponseStatus'
    type: object
...
  dto.ChapterResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      lessons:
        items:
          $ref: '#/definitions/dto.LessonResponse'
        type: array
      order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateLessonAssetRequest:
    properties:
      document_url:
        type: string
      duration:
        type: integer
      is_able_to_preview:
        type: boolean
      title:
        maxLength: 255
        minLength: 3
        type: string
      video_url:
        type: string
    required:
    - duration
    - title
    type: object
  dto.CreateLessonRequest:
    properties:
      chapter_title:
        maxLength: 255
        minLength: 3
        type: string
      lessons:
        items:
          $ref: '#/definitions/dto.CreateLessonAssetRequest'
        minItems: 1
        type: array
    required:
    - chapter_title
    - lessons
    type: object
  dto.CreateLessonsRequest:
    properties:
      chapters:
        items:
          $ref: '#/definitions/dto.CreateLessonRequest'
        minItems: 1
        type: array
    required:
    - chapters
    type: object
...
  dto.LessonResponse:
    properties:
      created_at:
        type: string
      document_url:
        type: string
      duration:
        type: integer
      id:
        type: string
      is_able_to_preview:
        type: boolean
      order:
        type: integer
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
      video_url:
        type: string
    type: object
...
  dto.ResponseStatus:
    properties:
      code:
        type: integer
      type:
        type: string
    type: object
...
  dto.UpdateChapterWithLessonsRequest:
    properties:
      id:
        description: Empty = new chapter, filled = update/keep existing
        type: string
      lessons:
        items:
          $ref: '#/definitions/dto.UpdateLessonInChapterRequest'
        minItems: 1
        type: array
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - lessons
    - title
    type: object
...
  dto.UpdateCourseWithChaptersRequest:
    properties:
      chapters:
        items:
          $ref: '#/definitions/dto.UpdateChapterWithLessonsRequest'
        minItems: 1
        type: array
    required:
    - chapters
    type: object
...
  dto.UpdateLessonInChapterRequest:
    properties:
      document_url:
        type: string
      duration:
        type: integer
      id:
        description: Empty = new lesson, filled = update/keep existing
        type: string
      is_able_to_preview:
        type: boolean
      title:
        maxLength: 255
        minLength: 3
        type: string
      video_url:
        type: string
    required:
    - duration
    - title
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/courses/{id}/lessons:
    get:
      consumes:
      - application/json
      description: Retrieve the full structure of a course including chapters and
        their associated lessons
      parameters:
      - description: Course ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ChapterResponse'
                  type: array
              type: object
        "400":
          description: Invalid course ID
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get all chapters and lessons of a course
      tags:
      - lessons
    post:
      consumes:
      - application/json
      description: Create one chapter and multiple lessons in a single request using
        presigned media URLs
      parameters:
      - description: Course ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Chapters and lessons create request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLessonsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ChapterResponse'
                  type: array
              type: object
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create a chapter with lessons
      tags:
      - lessons
    put:
      consumes:
      - application/json
      description: Update course structure with chapters and lessons (supports create,
        update, delete)
      parameters:
      - description: Course ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Course chapters and lessons update request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCourseWithChaptersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ChapterResponse'
                  type: array
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update course with all chapters and lessons
      tags:
      - lessons
...
  /api/v1/events/{event_id}/meeting/ws:
    get:
      description: 'WebSocket endpoint for event meeting room. Require `room_token`
        query and either `Authorization: Bearer <token>` header or `access_token`
        query.'
      parameters:
      - description: Event ID (UUID format)
        in: path
        name: event_id
        required: true
        type: string
      - description: Room token generated by event
        in: query
        name: room_token
        required: true
        type: string
      - description: JWT access token (for browser websocket fallback)
        in: query
        name: access_token
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Join event meeting websocket
      tags:
      - meeting
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"