# elearning/e-smart-learn/app/handler/meeting_handler_ws.go
package handler

import (
	"context"
	"fmt"
	"net/http"
	"strings"

	"elearning-api/apperror"
	"elearning-api/config"
	"elearning-api/service"
	"elearning-api/util"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/gorilla/websocket"
)

type MeetingHandler interface {
	JoinEventMeeting() gin.HandlerFunc
}

type meetingHandler struct {
	meetingService service.MeetingService
	meetingHub     *service.MeetingHub
	jwtConfig      *config.JWTConfig
	upgrader       websocket.Upgrader
}

func NewMeetingHandler(
	meetingService service.MeetingService,
	meetingHub *service.MeetingHub,
	jwtConfig *config.JWTConfig,
) MeetingHandler {
	return &meetingHandler{
		meetingService: meetingService,
		meetingHub:     meetingHub,
		jwtConfig:      jwtConfig,
		upgrader: websocket.Upgrader{
			ReadBufferSize:  1024,
			WriteBufferSize: 1024,
			CheckOrigin: func(r *http.Request) bool {
				return true
			},
		},
	}
}

// JoinEventMeeting godoc
// @Summary Join event meeting websocket
// @Description WebSocket endpoint for event meeting room. Require `room_token` query and either `Authorization: Bearer <token>` header or `access_token` query.
// @Tags meeting
// @Produce json
// @Param event_id path string true "Event ID (UUID format)"
// @Param room_token query string true "Room token generated by event"
// @Param access_token query string false "JWT access token (for browser websocket fallback)"
// @Success 101 {string} string "Switching Protocols"
// @Failure 400 {object} dto.ApiResponse "Invalid request"
// @Failure 401 {object} dto.ApiResponse "Unauthorized"
// @Failure 403 {object} dto.ApiResponse "Forbidden"
// @Failure 404 {object} dto.ApiResponse "Event not found"
// @Router /api/v1/events/{event_id}/meeting/ws [get]
func (h *meetingHandler) JoinEventMeeting() gin.HandlerFunc {
	return func(c *gin.Context) {
		eventID, err := uuid.Parse(c.Param("event_id"))
		if err != nil {
			_ = c.Error(apperror.NewBadRequestError("Invalid event ID format"))
			return
		}

		roomToken := strings.TrimSpace(c.Query("room_token"))
		if roomToken == "" {
			_ = c.Error(apperror.NewBadRequestError("room_token is required"))
			return
		}
		token := c.Query("access_token")
		if token == "" {
			_ = c.Error(apperror.NewBadRequestError("access_token is required"))
			return
		}
		userID, err := h.validateToken(c.Request.Context(), token)
		if err != nil {
			_ = c.Error(apperror.NewUnauthorizedError("Unauthorized: " + err.Error()))
			return
		}

		joinCtx, err := h.meetingService.ValidateJoin(c.Request.Context(), userID, eventID, roomToken)
		if err != nil {
			if err, ok := err.(*apperror.AppError); ok {
				_ = c.Error(err)
				return
			}
			_ = c.Error(apperror.NewInternalServerError("Unable to join meeting"))
			return
		}

		conn, err := h.upgrader.Upgrade(c.Writer, c.Request, nil)
		if err != nil {
			_ = c.Error(apperror.NewInternalServerError("Failed to upgrade connection"))
			return
		}

		h.meetingHub.HandleConnection(conn, joinCtx.RoomID, joinCtx.EventID, joinCtx.Participant)
	}
}

func (h *meetingHandler) validateToken(ctx context.Context, token string) (uuid.UUID, error) {
	claims, err := util.ValidateAccessToken(token, h.jwtConfig)
	if err != nil {
		util.LoggerFromContextWithLayer(ctx, util.LayerHandler).Warn("token validation error", "error", err)
		return uuid.Nil, err
	}

	userID, err := uuid.Parse(claims.UserID)
	if err != nil {
		return uuid.Nil, fmt.Errorf("invalid user ID in token: %w", err)
	}

	return userID, nil
}